// apps/web/src/pages/products/types/product.types.ts
import type { Category } from "./category.types";

export interface ProductBranchStock {
  branchId: number;
  branchName: string;
  quantity: number;
}

export interface Product {
  id: number;
  name: string;
  barcode: string;
  stockQuantity: number;
  branchStockQuantity?: number | null;
  companyStockQuantity?: number | null;
  branchStocks?: ProductBranchStock[];
  stockUnit: string;
  description?: string | null;
  salePrice: number;
  purchasePrice?: number | null;
  imageUrl?: string | null;
  isActive: boolean;
  parentCategoryId: number;
  subCategoryId?: number | null;
  parentCategory?: Category | null;
  subCategory?: Category | null;
}